common: avoid theoretical SIGHUP loop

+++ common/ECLogger.cpp
@@ -557,7 +557,8 @@ ECLogger_Pipe::~ECLogger_Pipe() {
 
 void ECLogger_Pipe::Reset() {
 	// send the log process HUP signal again
-	kill(m_childpid, SIGHUP);
+	if (m_childpid > 0)
+		kill(m_childpid, SIGHUP);
 }
 
 void ECLogger_Pipe::Log(unsigned int loglevel, const std::string &message) {
